Update strong params
