+++ app/controllers/users_controller.rb
@@ -9,6 +9,7 @@ class UsersController < ApplicationController
 
   def show
     @user = User.find(params[:id])
+    # require'pry';binding.pry
   end
 
   private
